pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bourdon"
version = "0.0.12"
description = "Cross-agent memory federation for human-AI collaboration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Ryan Davis", email = "ryandavispro1@gmail.com" },
]
maintainers = [
    { name = "RADLAB LLC" },
]
keywords = [
    "ai",
    "memory",
    "agent",
    "federation",
    "mcp",
    "rag",
    "cognition",
    "llm",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "jsonschema>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
# L2 episodic memory via UltraRAG / any MCP retriever.
# Optional because L0+L1 work standalone; L2 only activates when enabled.
ultrarag = [
    "fastmcp>=2.0",
]
# L6 federation MCP server. Required to run `python -m core.l6_server`
# and expose the agent-library via the Model Context Protocol.
server = [
    "fastmcp>=2.0",
]
# llama.cpp inference backend (adapters/llama_cpp_backend.py).
# Optional because not every Bourdon user runs local inference; the
# Protocol surface in core/inference_protocol.py has no httpx dependency.
llama-cpp = [
    "httpx>=0.27",
]

# Entry points for adapter plugins.
# Third-party adapters register here via their own pyproject.toml.
[project.entry-points."bourdon.adapters"]
claude-code = "adapters.claude_code:ClaudeCodeAdapter"
codex = "adapters.codex:CodexAdapter"
cursor = "adapters.cursor:CursorAdapter"
# clyde = "adapters.clyde:ClydeAdapter"     # native publisher lives in ~/.clyde/, not here
# clair = "adapters.clair:ClairAdapter"     # same pattern -- native publisher

[project.scripts]
bourdon = "cli.main:main"

[project.urls]
Homepage = "https://bourdon.ai"
Repository = "https://github.com/getbourdon/bourdon"
Issues = "https://github.com/getbourdon/bourdon/issues"
Documentation = "https://bourdon.ai/docs"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "adapters*", "cli*"]
exclude = ["tests*", "docs*", "spec*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: live tests requiring an external service (e.g. llama-server). Skipped in CI by default; run locally with `pytest -m integration`.",
]
